+++ .agents/scripts/lib/orchestration/epic-runner/checkpointer.js
@@ -11,39 +11,17 @@ import {
   findStructuredComment,
   upsertStructuredComment,
 } from '../ticketing.js';
+import { assertValidDeliverPhase } from './deliver-phases.js';
 
 export const EPIC_RUN_STATE_TYPE = 'epic-run-state';
 export const CHECKPOINT_SCHEMA_VERSION = 1;
 
-/**
- * Ordered list of `/epic-deliver` phases (Story #1155 / Epic #1142). The
- * checkpoint's `phase` field stores the **next phase to run**, so a
- * mid-flight crash during `code-review` resumes by reading
- * `phase === 'code-review'` and re-entering Phase D from the start.
- *
- * Phase tags are stable identifiers — downstream tooling (the contract
- * test in `tests/workflows/epic-deliver.test.js`) asserts the phase
- * advancement contract directly against this list.
- */
-export const DELIVER_PHASES = Object.freeze([
-  'prepare',
-  'wave-loop',
-  'close-validation',
-  'code-review',
-  'retro',
-  'finalize',
-]);
-
-/**
- * Pure: index of `phase` in `DELIVER_PHASES`. Returns `-1` for unknown
- * values (callers treat that as "start fresh"); `+Infinity` for the
- * terminal `'done'` sentinel.
- */
-export function phaseIndex(phase) {
-  if (phase === 'done') return Number.POSITIVE_INFINITY;
-  const idx = DELIVER_PHASES.indexOf(phase);
-  return idx;
-}
+// Re-export the phase enum + index helper so downstream importers
+// continue to use `checkpointer.js` as a single import target.
+export {
+  DELIVER_PHASES,
+  phaseIndex,
+} from './deliver-phases.js';
 
 export class Checkpointer {
   /**
@@ -133,12 +111,7 @@ export class Checkpointer {
    * @param {string} nextPhase - One of `DELIVER_PHASES` or `'done'`.
    */
   async setPhase(nextPhase) {
-    if (nextPhase !== 'done' && phaseIndex(nextPhase) < 0) {
-      throw new Error(
-        `Checkpointer.setPhase: invalid phase ${JSON.stringify(nextPhase)}. ` +
-          `Expected one of ${DELIVER_PHASES.join(', ')} or 'done'.`,
-      );
-    }
+    assertValidDeliverPhase(nextPhase);
     const existing = (await this.read()) ?? {};
     return this.write({ ...existing, phase: nextPhase });
   }
