+++ baselines/maintainability.json
@@ -18,10 +18,10 @@
   ".agents/scripts/drain-pending-cleanup.js": 106.069,
   ".agents/scripts/epic-close.js": 95.767,
   ".agents/scripts/epic-code-review.js": 84.268,
+  ".agents/scripts/epic-deliver-finalize.js": 85.829,
   ".agents/scripts/epic-deliver-prepare.js": 93.605,
   ".agents/scripts/epic-deliver-runner.js": 94.617,
   ".agents/scripts/epic-execute-record-wave.js": 86.848,
-  ".agents/scripts/epic-finalize.js": 89.011,
   ".agents/scripts/epic-plan-decompose.js": 86.294,
   ".agents/scripts/epic-plan-healthcheck.js": 98.816,
   ".agents/scripts/epic-plan-spec.js": 79.607,
@@ -94,6 +94,7 @@
   ".agents/scripts/lib/observability/perf-aggregator.js": 90.365,
   ".agents/scripts/lib/observability/signals-writer.js": 87.766,
   ".agents/scripts/lib/observability/tool-trace-hook.js": 94.506,
+  ".agents/scripts/lib/orchestration/code-review.js": 90.238,
   ".agents/scripts/lib/orchestration/concurrency.js": 106.379,
   ".agents/scripts/lib/orchestration/concurrent-task-resolver.js": 89.513,
   ".agents/scripts/lib/orchestration/context-hydration-engine.js": 92.569,
@@ -102,13 +103,15 @@
   ".agents/scripts/lib/orchestration/dispatch-engine.js": 98.293,
   ".agents/scripts/lib/orchestration/dispatch-pipeline.js": 93.83,
   ".agents/scripts/lib/orchestration/doc-reader.js": 102.728,
+  ".agents/scripts/lib/orchestration/epic-deliver-close-tail.js": 67.92,
   ".agents/scripts/lib/orchestration/epic-lifecycle-detector.js": 103.961,
   ".agents/scripts/lib/orchestration/epic-runner.js": 111.323,
   ".agents/scripts/lib/orchestration/epic-runner/blocker-handler.js": 91.735,
   ".agents/scripts/lib/orchestration/epic-runner/bookend-chainer.js": 89.316,
-  ".agents/scripts/lib/orchestration/epic-runner/checkpointer.js": 106.165,
+  ".agents/scripts/lib/orchestration/epic-runner/checkpointer.js": 106.565,
   ".agents/scripts/lib/orchestration/epic-runner/column-sync.js": 88.882,
   ".agents/scripts/lib/orchestration/epic-runner/commit-assertion.js": 97.501,
+  ".agents/scripts/lib/orchestration/epic-runner/deliver-phases.js": 118.393,
   ".agents/scripts/lib/orchestration/epic-runner/factory.js": 96.895,
   ".agents/scripts/lib/orchestration/epic-runner/phases/build-wave-dag.js": 105.841,
   ".agents/scripts/lib/orchestration/epic-runner/phases/finalize.js": 101.927,
@@ -139,6 +142,7 @@
   ".agents/scripts/lib/orchestration/reconciler.js": 106.792,
   ".agents/scripts/lib/orchestration/recut.js": 111.264,
   ".agents/scripts/lib/orchestration/retro-heuristics.js": 114.637,
+  ".agents/scripts/lib/orchestration/retro-runner.js": 90.07,
   ".agents/scripts/lib/orchestration/story-close-recovery.js": 85.963,
   ".agents/scripts/lib/orchestration/story-close/baseline-attribution-wiring.js": 82.703,
   ".agents/scripts/lib/orchestration/story-close/baseline-attribution.js": 93.072,
@@ -292,7 +296,6 @@
   "tests/epic-code-review.test.js": 104.492,
   "tests/epic-execute/epic-deliver-prepare.test.js": 97.923,
   "tests/epic-execute/epic-execute-record-wave.test.js": 101.537,
-  "tests/epic-execute/epic-finalize.test.js": 98.679,
   "tests/epic-plan-spec-drain.test.js": 102.552,
   "tests/epic-planner.test.js": 104.969,
   "tests/epic-runner/_build-ctx.js": 122.507,
@@ -387,6 +390,7 @@
   "tests/lib/observability/perf-aggregator.test.js": 94.132,
   "tests/lib/observability/signals-writer.test.js": 104.033,
   "tests/lib/observability/tool-trace-hook.test.js": 103.898,
+  "tests/lib/orchestration/code-review.test.js": 102.057,
   "tests/lib/orchestration/concurrent-task-resolver.test.js": 88.289,
   "tests/lib/orchestration/context.test.js": 121.009,
   "tests/lib/orchestration/dispatch-pipeline-runWorktreeGc.test.js": 109.936,
@@ -398,6 +402,7 @@
   "tests/lib/orchestration/phase-runner.test.js": 121.235,
   "tests/lib/orchestration/plan-runner/worktree-sweep.test.js": 94.608,
   "tests/lib/orchestration/post-merge-pipeline.test.js": 106.084,
+  "tests/lib/orchestration/retro-runner.test.js": 91.294,
   "tests/lib/orchestration/story-close/baseline-attribution-wiring.test.js": 100.587,
   "tests/lib/orchestration/story-close/baseline-attribution.test.js": 93.961,
   "tests/lib/orchestration/story-close/close-validation.test.js": 103.627,
@@ -474,6 +479,7 @@
   "tests/run-audit-suite-cli.test.js": 106.961,
   "tests/schemas/signal-schemas.test.js": 105.795,
   "tests/scripts/agents-bootstrap-github-protection.test.js": 98.807,
+  "tests/scripts/epic-deliver-finalize.test.js": 108.57,
   "tests/select-audits-cli.test.js": 109.859,
   "tests/shipped-configs-validate.test.js": 105.887,
   "tests/story-close-cd-out-guard.test.js": 100.748,
@@ -499,7 +505,7 @@
   "tests/workflows/epic-close.aux-concurrency.test.js": 107.413,
   "tests/workflows/epic-close.phase-6.test.js": 113.261,
   "tests/workflows/epic-close.phase-7.test.js": 99.13,
-  "tests/workflows/epic-close.smoke.test.js": 121.097,
+  "tests/workflows/epic-deliver.test.js": 104.975,
   "tests/workflows/epic-execute-record-wave-cli.test.js": 110.431,
   "tests/workflows/epic-plan-ideation.test.js": 111.497,
   "tests/workflows/epic-planner-cli.test.js": 111.339,
